dealt with the few kinda valid compiler warnings that are not caused bei sfml/box2d/tileson files. fallback return for asset manager needed still
(check if joystick-controls still work, changes marked with comments)

+++ src/InputManager.cpp
@@ -130,7 +130,9 @@ bool InputManager::isJoystickButtonBound(JoystickButton button)
 
 void InputManager::processJoystickMoved(const sf::Event& event)
 {
-    int squaredPos = event.joystickMove.position * event.joystickMove.position;
+    // changed squared pos from int to float cause compiler warnings
+    // pls test if that causes any issues
+    float squaredPos = event.joystickMove.position * event.joystickMove.position;
     bool notZero= squaredPos > JoystickMap::ZERO_SQAURED;
 
     if (notZero)
